requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Streamlit GUI
streamlit==1.30.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Cache
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Multi-Agent & AI
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1

# Local LLM Support
ollama==0.1.6

# Browser Automation
playwright==1.41.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# Monitoring & Observability
prometheus-client==0.19.0
python-json-logger==2.0.7

# Task Queue
celery==5.3.6
flower==2.0.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.3

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-pytest==0.21.3
faker==22.0.0

# Development
black==24.1.0
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
